% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_symb.R
\name{cr_symb}
\alias{cr_symb}
\title{A function to create symbols}
\usage{
cr_symb(k = k, m = m)
}
\arguments{
\item{k}{number of categories}

\item{m}{length of the m-surrounding}
}
\value{
A list with two types of symbols. Permutation and Combinations-totals
 \tabular{ll}{
    \code{p_symb} \tab Matrix with symbols (permutations)\cr
    \code{c_symb} \tab  Matrix with symbols (combinations)\cr
    }
}
\description{
This function obtains the set of symbols to get the Q-statistic
}
\details{
...
}
\examples{

# Example 1: Obtain symbols for k=2 classes and m-surroundings of size 5
symb25 <- cr_symb(2,5)
symb25$p_symb # Permutations symbols
symb25$c_symb # Combinations-totals symbols
}
\references{
\itemize{
    \item Ruiz M, López FA, A Páez. (2010). \emph{Testing for spatial association of qualitative
    data using symbolic dynamics}. Journal of Geographical Systems. 12 (3) 281-309
  }
}
\seealso{
\code{\link{m.surround}}
}
\author{
\tabular{ll}{
  Fernando López  \tab \email{fernando.lopez@upct.es} \cr
  Román Mínguez  \tab \email{roman.minguez@uclm.es} \cr
  Antonio Páez \tab \email{paezha@gmail.com} \cr
  Manuel Ruiz \tab \email{manuel.ruiz@upct.es} \cr
  }
}
