% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{FastFood.sf}
\alias{FastFood.sf}
\title{Selection of fast food restaurants in Toronto}
\format{
A simple features object with 614 rows and 3 variables:

\describe{
  \item{ID}{Unique identifier of record.}
  \item{Class}{Factor with 3 types of fast food restaurants:
  [P]izza, [S]andwich, and [H]amburger}
  \item{geometry}{Geometry of simple features.}
}
}
\source{
Ruiz et al. (2010)
\url{https://link.springer.com/article/10.1007/s10109-009-0100-1}
}
\usage{
data(FastFood.sf)
}
\description{
A simple feature (sf) dataframe containing the locations of a selection of fast food restaurants
in the city of Toronto, Canada (data are from 2008). The data are projected using EPSG: 26917 (WGS 84/UTM Zone 17N).
}
\references{
\itemize{
    \item Ruiz M, López FA, A Páez. (2010). \emph{Testing for spatial association of qualitative
    data using symbolic dynamics}. Journal of Geographical Systems. 12 (3) 281-309
  }
}
\keyword{datasets}
