% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_scantest.R
\name{methods_scantest}
\alias{methods_scantest}
\alias{plot.scantest}
\alias{summary.scantest}
\title{Methods for class scantest}
\usage{
\method{plot}{scantest}(x, ..., sf = NULL, coor = NULL)

\method{summary}{scantest}(object, ...)
}
\arguments{
\item{x}{is a \code{scantest} object created by \code{\link{scan.test}} for \code{plot()} method.}

\item{...}{further arguments passed to or from other methods.}

\item{sf}{optional argument for \code{plot()} method to include a sf object (default = NULL)}

\item{coor}{optional argument for \code{plot()} method to include coordinates of points (default = NULL)}

\item{object}{a \code{scantest} object created by \code{\link{scan.test}} for \code{summary()} method.}
}
\value{
No return value, called for side effects
}
\description{
The \code{plot()} function allows the user plot the significant cluster(s).\cr
 \code{\link{summary}} list information about the most likelihood cluster.\cr
}
\references{
\itemize{
    \item Ruiz, M., López, F., and Páez, A. (2021).
    A test for global and local homogeneity of categorical data based on spatial runs.
      \emph{Working paper}.
  }
}
\author{
\tabular{ll}{
  Fernando López  \tab \email{fernando.lopez@upct.es} \cr
  Román Mínguez  \tab \email{roman.minguez@uclm.es} \cr
  Antonio Páez \tab \email{paezha@gmail.com} \cr
  Manuel Ruiz \tab \email{manuel.ruiz@upct.es} \cr
  }
}
