% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_spqdata.R
\name{methods_spqtest}
\alias{methods_spqtest}
\alias{plot.spqtest}
\title{Methods for class spqtest}
\usage{
\method{plot}{spqtest}(x, ..., ci = 0.95)
}
\arguments{
\item{x}{a \code{spqtest} object created by \code{\link{Q.test}}.}

\item{...}{further arguments passed to or from other methods}

\item{ci}{confidence level for the intervals in \code{plot} method.
Default \code{ci = 0.95}}
}
\value{
This functions does not return any value
}
\description{
The \code{plot()} function allows the user obtain the plot of relative frequency
of each symbol (standard and equivalent) with the confidence interval.
 The \code{print()} function is used to get a list with the values of Q test
 for standard symbolization based on permutations and for equivalent symbolization
 based on combinations. \code{summary()} print a table with the output of the Q test.
}
\references{
\itemize{
    \item Ruiz, M., López, F., and Páez, A. (2021).
    A test for global and local homogeneity of categorical data based on spatial runs.
      \emph{working paper}.
  }
}
\author{
\tabular{ll}{
  Fernando López  \tab \email{fernando.lopez@upct.es} \cr
  Román Mínguez  \tab \email{roman.minguez@uclm.es} \cr
  Antonio Páez \tab \email{paezha@gmail.com} \cr
  Manuel Ruiz \tab \email{manuel.ruiz@upct.es} \cr
  }
}
