% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.spqtest.R
\name{print.summary.spqtest}
\alias{print.summary.spqtest}
\title{Print method for objects of class summary.spqtest.}
\usage{
\method{print}{summary.spqtest}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{object of class \emph{summary.spqtest}.}

\item{digits}{number of digits to show in printed tables.
Default: max(3L, getOption("digits") - 3L).}

\item{...}{further arguments passed to or from other methods.}
}
\value{
No return value, called for side effects
}
\description{
Print method for objects of class summary.spqtest.
}
\seealso{
\code{\link{summary.spqtest}}.
}
\author{
\tabular{ll}{
  Fernando López  \tab \email{fernando.lopez@upct.es} \cr
  Román Mínguez  \tab \email{roman.minguez@uclm.es} \cr
  Antonio Páez \tab \email{paezha@gmail.com} \cr
  Manuel Ruiz \tab \email{manuel.ruiz@upct.es} \cr
  }
}
