% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spqdep-package.R
\docType{package}
\name{spqdep-package}
\alias{spqdep}
\alias{spqdep-package}
\title{spqdep: Testing for Spatial Independence of Cross-Sectional Qualitative Data}
\description{
Testing for Spatial Dependence of Qualitative Data in Cross Section. The list of functions includes join-count tests, Q test, spatial scan test, similarity test and spatial runs test. The methodology of these models can be found in \doi{10.1007/s10109-009-0100-1} and \doi{10.1080/13658816.2011.586327}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://f8l5h9.github.io/spqdep/}
}

}
\author{
\strong{Maintainer}: Fernando Lopez \email{fernando.lopez@upct.es} (\href{https://orcid.org/0000-0002-5397-9748}{ORCID})

Authors:
\itemize{
  \item Roman Minguez \email{roman.minguez@uclm.es} (\href{https://orcid.org/0000-0002-0490-3181}{ORCID})
  \item Antonio Paez \email{paezha@gmail.com} (\href{https://orcid.org/0000-0001-6912-9919}{ORCID})
  \item Manuel Ruiz \email{manuel.ruiz@upct.es} (\href{https://orcid.org/0000-0001-9228-6410}{ORCID})
}

}
\keyword{internal}
