% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Boots.sf}
\alias{Boots.sf}
\title{Boots.sf.}
\format{
A simple features dataframe with 256 rows and 1 column:
\describe{
  \item{BW}{A factor with two levels: Black and White}
}
}
\source{
Boots, B. (2003) \url{https://link.springer.com/content/pdf/10.1007/s10109-003-0110-3.pdf}
}
\usage{
data(Boots.sf)
}
\description{
A simple features object square regular lattice 16x16.
}
\examples{
 data(Boots.sf)
 summary(Boots.sf)
 plot(Boots.sf)
}
\references{
\itemize{
    \item Boots, B. (2003). Developing local measures of spatial association
    for categorical data. Journal of Geographical Systems, 5(2), 139-160.
    \item Upton G, Fingleton B (1985) Spatial data analysis by example.
    Volume 1: Point pattern and quantitative data.
    John Wiley & Sons, Chichester
  }
}
\keyword{datasets}
